+++ plugin-build/src/test/kotlin/io/sentry/android/gradle/integration/SentryPluginTest.kt
@@ -899,6 +899,38 @@ class SentryPluginTest :
         assertEquals(uuid1, uuid2)
     }
 
+    @Test
+    fun `copyFlutterAssetsDebug is wired up`() {
+        assumeThat(
+            "We only wire up the copyFlutterAssets task " +
+                "if the transform API (AGP >= 7.4.0) is used",
+            SemVer.parse(androidGradlePluginVersion) >= AgpVersions.VERSION_7_4_0,
+            `is`(true)
+        )
+
+        // when a flutter project is detected
+        appBuildFile.appendText(
+            // language=Groovy
+            """
+            tasks.register("copyFlutterAssetsDebug") {
+                doFirst {
+                    println("Hello World")
+                }
+            }
+            """.trimIndent()
+        )
+
+        // then InjectSentryMetaPropertiesIntoAssetsTask should depend on it
+        // and thus copyFlutterAssetsDebug should be executed as part of assembleDebug
+        val build = runner.withArguments(":app:assembleDebug").build()
+
+        assertEquals(
+            TaskOutcome.SUCCESS,
+            build.task(":app:copyFlutterAssetsDebug")?.outcome,
+            build.output
+        )
+    }
+
     private fun applyUploadNativeSymbols() {
         appBuildFile.appendText(
             // language=Groovy